% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsa.R
\name{fsa_ate}
\alias{fsa_ate}
\title{Your title goes here: example: sensitivity boot}
\usage{
fsa_ate(X, Z, Y, eps1_list, eps0_list, n_boot = 500, truncpscore = c(0, 1))
}
\arguments{
\item{X}{an R array}

\item{Y}{an R array}
}
\value{
A number.
}
\description{
Your description goes here: example: Creates a plot of the crayon colors in \code{\link{brocolors}}
}
\examples{
fsa_ate()
}
