% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsa.R
\name{fsa_point}
\alias{fsa_point}
\title{Your title goes here: example: sensitivity point analysis}
\usage{
fsa_point(X, Z, Y, epsilon1, epsilon0, truncpscore = c(0, 1))
}
\arguments{
\item{X}{an R array}

\item{Y}{an R array}
}
\value{
A number.
}
\description{
Your description goes here: example: Creates a plot of the crayon colors in \code{\link{brocolors}}
}
\examples{
fsa_point()
}
