% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsa.R
\name{plot_contour}
\alias{plot_contour}
\title{Your title goes here: example: sensitivity boot}
\usage{
plot_contour(ate_result, caption = "", estimator = "dr", value = "est")
}
\arguments{
\item{X}{an R array}

\item{Y}{an R array}
}
\value{
A number.
}
\description{
Your description goes here: example: Creates a plot of the crayon colors in \code{\link{brocolors}}
}
\examples{
plot_contour()
}
